pyproject.toml
[project]
name = "microtrade-fwf"
version = "0.2.9"
description = "Turn zipped fixed-width (FWF) trade microdata into Hive-partitioned Parquet datasets."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "microtrade contributors" }]
keywords = [
    "trade",
    "microdata",
    "fwf",
    "fixed-width",
    "parquet",
    "hive-partitioned",
    "data-pipeline",
    "statcan",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "polars[excel]>=1.20",
    "pyarrow>=17",
    "typer>=0.12",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "loguru>=0.7",
]

[project.urls]
Homepage = "https://github.com/twedl/microtrade"
Repository = "https://github.com/twedl/microtrade"
Issues = "https://github.com/twedl/microtrade/issues"
Changelog = "https://github.com/twedl/microtrade/blob/main/CHANGELOG.md"

[project.scripts]
microtrade = "microtrade.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/microtrade"]
# Reference YAML specs are committed at `src/microtrade/specs/` for use as
# test fixtures and as starting-point examples in-repo. They are not the
# runtime contract for installed users (each consumer writes their own via
# `microtrade import-spec`), so keep them out of the wheel.
exclude = ["src/microtrade/specs"]

[tool.hatch.build.targets.sdist]
exclude = ["src/microtrade/specs"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "types-PyYAML>=6.0",
    "types-tqdm>=4.66",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
# Typer's entire API relies on function-call defaults (B008).
"src/microtrade/cli.py" = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src/microtrade"]

[[tool.mypy.overrides]]
module = ["pyarrow.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra --cov=microtrade --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
source = ["microtrade"]
branch = true
